requirements.txt
fastapi[standard]
pytz
python-dotenv
aiohttp
python-statemachine[diagrams]